Changed the data structure output to something more linear. for each detector we have: sim hits with: channel num, edep (if relevent), npe, time, t_lead, t_trail (if relevent); each are stored in separate vectors, all the same size for a given detector (of course, each detector has its own set of vectors).
hits with channel number, data word, and unencoded ADC and TDC value, whichever is relevent (the other value being set to -1);

These data structures are part of TSBSSimEvent, and are set as the output in TSBSSimDigitizer (and filled in the detector class). I had implemented it on TSBSSimScint for the example.

On the side, I also added some storage containers for the MC info for the TMPTSignal class.

Lats note: the elements of the previous output are still there and not commented out until the new container outputs are not filled properly for all detecto types.